refactor(intern): refactored the intern code to have an average LOC of 500 lines . resolved assigned_to ID mismatch in task assignment
