+++ src/app/(dashboard)/dashboard/intern/tasks/intern-task-client.tsx
@@ -4,7 +4,6 @@ import { Clock, Search } from "lucide-react";
 import { useState } from "react";
 import { toast } from "sonner";
 import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
 import {
   Card,
   CardContent,
@@ -12,14 +11,6 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
 import {
   Select,
@@ -39,6 +30,8 @@ import {
   getTaskKarma,
 } from "@/features/intern/utils/intern-helpers";
 import { useDebounce } from "@/hooks/use-debounce";
+import { SubmitDeliverableDialog } from "./components/submit-deliverable-dialog";
+import { TaskDetailDialog } from "./components/task-detail-dialog";
 
 const getComplexityColor = (complexity: string) => {
   switch (complexity) {
@@ -115,7 +108,6 @@ export function InternTasksPageClient() {
           setIsSubmitLinkOpen(false);
           setCompletingTaskId(null);
           setOutputLink("");
-          // If the selected task in the detailed dialog is the one that was completed, update its status
           if (selectedTask && selectedTask.id === completingTaskId) {
             setSelectedTask((prev) =>
               prev
@@ -312,284 +304,26 @@ export function InternTasksPageClient() {
         </div>
       )}
 
-      {/* Detailed Task View Dialog */}
-      <Dialog
-        open={!!selectedTask}
-        onOpenChange={(o) => !o && setSelectedTask(null)}
-      >
-        <DialogContent className="w-full max-w-[calc(100%-2rem)] sm:max-w-2xl border-border/40 bg-card/95 backdrop-blur-2xl shadow-2xl">
-          {selectedTask && (
-            <>
-              <DialogHeader className="pt-3 sm:pt-0 pr-8 pb-4 border-b border-border/20">
-                <div className="flex flex-col sm:flex-row sm:items-center justify-between gap-4">
-                  <DialogTitle className="text-2xl font-black uppercase tracking-tight text-foreground break-all">
-                    {selectedTask.title}
-                  </DialogTitle>
-                  <div className="flex flex-wrap items-center gap-2 shrink-0">
-                    <Badge
-                      variant="outline"
-                      className={`text-[9px] uppercase font-black tracking-widest ${getComplexityColor(selectedTask.complexity)}`}
-                    >
-                      {selectedTask.complexity}
-                    </Badge>
-                    {selectedTask.team && (
-                      <Badge
-                        variant="default"
-                        className="text-[9px] font-black rounded-md uppercase tracking-wider px-2 py-0.5"
-                      >
-                        {selectedTask.team}
-                      </Badge>
-                    )}
-                    <Badge
-                      variant="outline"
-                      className="bg-success/10 text-success border-success/20 text-[9px] font-black rounded-md uppercase tracking-wider px-2 py-0.5"
-                    >
-                      {getTaskBaseKarma(selectedTask)} Karma
-                    </Badge>
-                    <Badge
-                      variant="outline"
-                      className="bg-brand-purple/10 text-brand-purple border-brand-purple/20 text-[9px] font-black rounded-md uppercase tracking-wider px-2 py-0.5"
-                    >
-                      {getTaskKarma(selectedTask)} Pts
-                    </Badge>
-                    <Badge
-                      variant="default"
-                      className="text-[9px] font-black rounded-md uppercase tracking-wider px-2 py-0.5"
-                    >
-                      {selectedTask.category}
-                    </Badge>
-                  </div>
-                </div>
-                <DialogDescription className="hidden">
-                  Task details and description view
-                </DialogDescription>
-              </DialogHeader>
+      <TaskDetailDialog
+        task={selectedTask}
+        onClose={() => setSelectedTask(null)}
+        onStatusChange={handleStatusChange}
+        onTaskUpdate={(updated) => setSelectedTask(updated)}
+      />
 
-              <div className="overflow-y-auto max-h-[calc(80vh-10rem)] space-y-6 pt-4 pr-1">
-                {/* Description */}
-                <div className="space-y-2">
-                  <h4 className="text-[10px] font-black uppercase tracking-widest text-muted-foreground">
-                    Description
-                  </h4>
-                  <div className="p-4 rounded-xl border border-border/40 bg-background/40 max-h-[220px] overflow-y-auto w-full min-w-0">
-                    <p className="text-sm font-medium text-foreground whitespace-pre-wrap leading-relaxed break-all">
-                      {selectedTask.description || "No description provided."}
-                    </p>
-                  </div>
-                </div>
-
-                {/* Grid info */}
-                <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
-                  <div className="space-y-1 bg-background/25 border border-border/20 p-3 rounded-xl">
-                    <p className="text-[9px] font-black uppercase tracking-widest text-muted-foreground/60 flex items-center gap-1">
-                      <Clock className="w-3 h-3 text-muted-foreground/50" />
-                      Deadline
-                    </p>
-                    <p className="text-sm font-bold text-foreground">
-                      {selectedTask.deadline
-                        ? new Date(selectedTask.deadline).toLocaleDateString(
-                            undefined,
-                            {
-                              month: "short",
-                              day: "numeric",
-                              year: "numeric",
-                            },
-                          )
-                        : "No deadline"}
-                    </p>
-                  </div>
-                  <div className="space-y-1 bg-background/25 border border-border/20 p-3 rounded-xl">
-                    <p className="text-[9px] font-black uppercase tracking-widest text-muted-foreground/60">
-                      Created By
-                    </p>
-                    <p className="text-sm font-bold text-foreground">
-                      {selectedTask.created_by_name ||
-                        selectedTask.created_by ||
-                        "Admin/Mentor"}
-                    </p>
-                  </div>
-                  <div className="space-y-1 bg-background/25 border border-border/20 p-3 rounded-xl">
-                    <p className="text-[9px] font-black uppercase tracking-widest text-muted-foreground/60">
-                      Karma
-                    </p>
-                    <p className="text-sm font-bold text-foreground">
-                      {getTaskBaseKarma(selectedTask)}
-                    </p>
-                  </div>
-                  <div className="space-y-1 bg-background/25 border border-border/20 p-3 rounded-xl">
-                    <p className="text-[9px] font-black uppercase tracking-widest text-muted-foreground/60">
-                      Points
-                    </p>
-                    <p className="text-sm font-bold text-foreground">
-                      {getTaskKarma(selectedTask)}
-                    </p>
-                  </div>
-                  {selectedTask.status === "COMPLETED" &&
-                    selectedTask.output_link && (
-                      <div className="space-y-1 bg-background/25 border border-border/20 p-3 rounded-xl col-span-1 sm:col-span-2">
-                        <p className="text-[9px] font-black uppercase tracking-widest text-muted-foreground/60">
-                          Submission Link
-                        </p>
-                        <a
-                          href={selectedTask.output_link}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          className="text-xs font-bold text-brand-blue hover:underline break-all"
-                        >
-                          {selectedTask.output_link}
-                        </a>
-                      </div>
-                    )}
-                </div>
-
-                {/* Status Updater inside modal */}
-                <div className="flex flex-col sm:flex-row items-start sm:items-center justify-between gap-4 p-4 rounded-xl border border-border/40 bg-brand-blue/5">
-                  <div>
-                    <h4 className="text-xs font-black uppercase tracking-wider text-foreground">
-                      Task Status
-                    </h4>
-                    <p className="text-[10px] font-bold text-muted-foreground uppercase tracking-wide mt-0.5">
-                      Update your work progress on this task
-                    </p>
-                  </div>
-
-                  <div className="flex items-center gap-3 w-full sm:w-auto">
-                    <Select
-                      value={
-                        [
-                          "WAITING_FOR_REVIEW",
-                          "IN_PROGRESS",
-                          "COMPLETED",
-                          "ON_HOLD",
-                        ].includes(selectedTask.status)
-                          ? selectedTask.status
-                          : undefined
-                      }
-                      onValueChange={(val) => {
-                        handleStatusChange(
-                          selectedTask.id,
-                          val as TInternTask["status"],
-                        );
-                        if (val !== "COMPLETED") {
-                          setSelectedTask((prev) =>
-                            prev
-                              ? {
-                                  ...prev,
-                                  status: val as TInternTask["status"],
-                                }
-                              : null,
-                          );
-                        }
-                      }}
-                      disabled={updateStatusMutation.isPending}
-                    >
-                      <SelectTrigger className="h-9 font-black uppercase text-[10px] tracking-widest w-fit min-w-[130px] px-3 border-border/50 bg-background/50 rounded-lg">
-                        <SelectValue placeholder="TODO" />
-                      </SelectTrigger>
-                      <SelectContent
-                        position="popper"
-                        className="bg-card font-bold border-border/60"
-                      >
-                        <SelectItem
-                          value="WAITING_FOR_REVIEW"
-                          className="uppercase text-[9px]"
-                        >
-                          Waiting for Review
-                        </SelectItem>
-                        <SelectItem
-                          value="IN_PROGRESS"
-                          className="uppercase text-[9px]"
-                        >
-                          In Progress
-                        </SelectItem>
-                        <SelectItem
-                          value="COMPLETED"
-                          className="uppercase text-[9px]"
-                        >
-                          Completed
-                        </SelectItem>
-                        <SelectItem
-                          value="ON_HOLD"
-                          className="uppercase text-[9px]"
-                        >
-                          On Hold
-                        </SelectItem>
-                      </SelectContent>
-                    </Select>
-                  </div>
-                </div>
-              </div>
-            </>
-          )}
-        </DialogContent>
-      </Dialog>
-
-      {/* Submit deliverables link dialog */}
-      <Dialog open={isSubmitLinkOpen} onOpenChange={setIsSubmitLinkOpen}>
-        <DialogContent className="w-full max-w-[calc(100%-2rem)] sm:max-w-md border-border/40 bg-card/95 backdrop-blur-2xl shadow-2xl">
-          <DialogHeader className="pb-4 border-b border-border/20">
-            <DialogTitle className="text-xl font-black uppercase tracking-tight text-foreground">
-              Submit Task Deliverables
-            </DialogTitle>
-            <DialogDescription className="text-xs font-semibold text-muted-foreground mt-1">
-              Please provide the submission link (e.g., GitHub PR, Figma link,
-              document) to mark this task as completed.
-            </DialogDescription>
-          </DialogHeader>
-
-          <div className="space-y-4 pt-4">
-            <div className="space-y-1.5">
-              <label
-                htmlFor="output-link-input"
-                className="text-[10px] font-black uppercase tracking-widest text-muted-foreground"
-              >
-                Submission URL / Output Link
-              </label>
-              <Input
-                id="output-link-input"
-                type="url"
-                required
-                pattern="https?://.+"
-                placeholder="https://github.com/... or similar"
-                value={outputLink}
-                onChange={(e) => setOutputLink(e.target.value)}
-                className="bg-background/50 border-border/50 font-medium focus-visible:ring-brand-blue"
-              />
-              <p className="text-[10px] text-muted-foreground font-medium">
-                Must be a valid URL starting with{" "}
-                <span className="font-black">https://</span>
-              </p>
-            </div>
-          </div>
-
-          <DialogFooter className="border-t border-border/20 pt-4 flex items-center justify-end gap-2 mt-4">
-            <Button
-              type="button"
-              variant="outline"
-              size="default"
-              onClick={() => {
-                setIsSubmitLinkOpen(false);
-                setCompletingTaskId(null);
-                setOutputLink("");
-              }}
-              className="text-xs uppercase tracking-widest"
-            >
-              Cancel
-            </Button>
-            <Button
-              type="button"
-              size="default"
-              onClick={handleConfirmComplete}
-              disabled={updateStatusMutation.isPending}
-              className="text-xs uppercase tracking-widest"
-            >
-              {updateStatusMutation.isPending
-                ? "Submitting..."
-                : "Complete Task"}
-            </Button>
-          </DialogFooter>
-        </DialogContent>
-      </Dialog>
+      <SubmitDeliverableDialog
+        open={isSubmitLinkOpen}
+        outputLink={outputLink}
+        isPending={updateStatusMutation.isPending}
+        onOpenChange={setIsSubmitLinkOpen}
+        onOutputLinkChange={setOutputLink}
+        onCancel={() => {
+          setIsSubmitLinkOpen(false);
+          setCompletingTaskId(null);
+          setOutputLink("");
+        }}
+        onConfirm={handleConfirmComplete}
+      />
     </div>
   );
 }
