+++ src/features/intern/components/index.ts
@@ -6,6 +6,12 @@
  * Barrel export for all intern-form components.
  */
 
+export type { UnifiedActivity } from "./activity-detail-dialog";
+export {
+  ActivityDetailDialog,
+  getActivityStatusBadge,
+} from "./activity-detail-dialog";
+export { QuestLog } from "./activity-log";
 export { QuestLogHistory } from "./activity-log-history";
 export { LeaveFormDialog } from "./leave-form-dialog";
 export { WeeklyReviewForm } from "./weekly-review-form";
